+++ src/int_on_evm/evm/initialize_evm_core.rs
@@ -50,7 +50,7 @@ use crate::{
 pub fn maybe_initialize_evm_core<D: DatabaseInterface>(
     db: D,
     block_json: &str,
-    chain_id: u8,
+    chain_id: u64,
     gas_price: u64,
     confs: u64,
 ) -> Result<String> {
